Tweak switcher to not scale icons down
This PR fixes so the new block icons are not scaled down, and therefore blurry, in the switcher.

Additionally it removes the gray background color from the "block chip". This is in response to similar changes elswhere in the UI.

It also adds a little dropdown arrow on hover, but not sure that adds much value.